Implemented Checkout and return procedures for a book.
Book can only be checked out if quantity for rent is greather than 0
implemented a stats that records the number of times a book has been checkedout
implemented a return feature
Person who checked out a book can only return it .
A book can not be returned if not checked out.
updating book catalalogue quantity for rent based on checkout and return.